Move CJK command path regression to Swift Testing
